Refactor move API/RSS classes to http NS and extend Output
